+++ ios/Classes/CountlyiOS/Countly.m
@@ -187,13 +187,19 @@ NSString* previousEventID;
     }
 
 #if (TARGET_OS_IOS || TARGET_OS_TV)
-    if ([config.features containsObject:CLYAutoViewTracking])
+    if (config.enableAutomaticViewTracking || [config.features containsObject:CLYAutoViewTracking])
     {
-        CountlyViewTracking.sharedInstance.isEnabledOnInitialConfig = YES;
-        [CountlyViewTracking.sharedInstance startAutoViewTracking];
+        // Print deprecation flag for feature
+        CountlyViewTrackingInternal.sharedInstance.isEnabledOnInitialConfig = YES;
+        [CountlyViewTrackingInternal.sharedInstance startAutoViewTracking];
+    }
+    if(config.automaticViewTrackingExclusionList) {
+        [CountlyViewTrackingInternal.sharedInstance addAutoViewTrackingExclutionList:config.automaticViewTrackingExclusionList];
     }
 #endif
-
+    if(config.globalViewSegmentation) {
+        [CountlyViewTrackingInternal.sharedInstance setGlobalViewSegmentation:config.globalViewSegmentation];
+    }
     timer = [NSTimer timerWithTimeInterval:config.updateSessionPeriod target:self selector:@selector(onTimer:) userInfo:nil repeats:YES];
     [NSRunLoop.mainRunLoop addTimer:timer forMode:NSRunLoopCommonModes];
 
@@ -213,7 +219,10 @@ NSString* previousEventID;
 
     [CountlyConnectionManager.sharedInstance proceedOnQueue];
 
-    if (config.consents)
+    //TODO: Should move at the top after checking the the edge cases of current implementation
+    if(config.enableAllConsents)
+        [self giveAllConsents];
+    else if (config.consents)
         [self giveConsentForFeatures:config.consents];
 
     if (config.campaignType && config.campaignData)
@@ -257,7 +266,7 @@ NSString* previousEventID;
     if (!CountlyCommon.sharedInstance.manualSessionHandling)
         [CountlyConnectionManager.sharedInstance endSession];
 
-    [CountlyViewTracking.sharedInstance pauseView];
+    [CountlyViewTrackingInternal.sharedInstance applicationDidEnterBackground];
 
     [CountlyPersistency.sharedInstance saveToFile];
 }
@@ -285,7 +294,7 @@ NSString* previousEventID;
     if (!CountlyCommon.sharedInstance.manualSessionHandling)
         [CountlyConnectionManager.sharedInstance beginSession];
 
-    [CountlyViewTracking.sharedInstance resumeView];
+    [CountlyViewTrackingInternal.sharedInstance applicationWillEnterForeground];
 
     isSuspended = NO;
 }
@@ -308,7 +317,7 @@ NSString* previousEventID;
 
     CountlyConnectionManager.sharedInstance.isTerminating = YES;
 
-    [CountlyViewTracking.sharedInstance endView];
+    [CountlyViewTrackingInternal.sharedInstance applicationWillTerminate];
 
     [CountlyConnectionManager.sharedInstance sendEvents];
 
@@ -569,7 +578,14 @@ NSString* previousEventID;
 {
     CLY_LOG_I(@"%s", __FUNCTION__);
 
-    [CountlyConsentManager.sharedInstance giveConsentForAllFeatures];
+    [CountlyConsentManager.sharedInstance giveAllConsents];
+}
+
+- (void)giveAllConsents
+{
+    CLY_LOG_I(@"%s", __FUNCTION__);
+    
+    [CountlyConsentManager.sharedInstance giveAllConsents];
 }
 
 - (void)cancelConsentForFeature:(NSString *)featureName
@@ -700,11 +716,11 @@ NSString* previousEventID;
 
     if ([key isEqualToString:kCountlyReservedEventView])
     {
-        event.PVID = CountlyViewTracking.sharedInstance.previousViewID ?: @"";
+        event.PVID = CountlyViewTrackingInternal.sharedInstance.previousViewID ?: @"";
     }
     else
     {
-        event.CVID = CountlyViewTracking.sharedInstance.currentViewID ?: @"";
+        event.CVID = CountlyViewTrackingInternal.sharedInstance.currentViewID ?: @"";
     }
 
     // Check if the event is a reserved event
@@ -953,14 +969,14 @@ NSString* previousEventID;
 {
     CLY_LOG_I(@"%s %@", __FUNCTION__, viewName);
 
-    [CountlyViewTracking.sharedInstance startView:viewName customSegmentation:nil];
+    [CountlyViewTrackingInternal.sharedInstance startAutoStoppedView:viewName segmentation:nil];
 }
 
 - (void)recordView:(NSString *)viewName segmentation:(NSDictionary *)segmentation
 {
     CLY_LOG_I(@"%s %@ %@", __FUNCTION__, viewName, segmentation);
 
-    [CountlyViewTracking.sharedInstance startView:viewName customSegmentation:segmentation];
+    [CountlyViewTrackingInternal.sharedInstance startAutoStoppedView:viewName segmentation:segmentation];
 }
 
 #if (TARGET_OS_IOS || TARGET_OS_TV)
@@ -968,31 +984,35 @@ NSString* previousEventID;
 {
     CLY_LOG_I(@"%s %@", __FUNCTION__, exception);
 
-    [CountlyViewTracking.sharedInstance addExceptionForAutoViewTracking:exception.copy];
+    [CountlyViewTrackingInternal.sharedInstance addExceptionForAutoViewTracking:exception.copy];
 }
 
 - (void)removeExceptionForAutoViewTracking:(NSString *)exception
 {
     CLY_LOG_I(@"%s %@", __FUNCTION__, exception);
 
-    [CountlyViewTracking.sharedInstance removeExceptionForAutoViewTracking:exception.copy];
+    [CountlyViewTrackingInternal.sharedInstance removeExceptionForAutoViewTracking:exception.copy];
 }
 
 - (void)setIsAutoViewTrackingActive:(BOOL)isAutoViewTrackingActive
 {
     CLY_LOG_I(@"%s %d", __FUNCTION__, isAutoViewTrackingActive);
 
-    CountlyViewTracking.sharedInstance.isAutoViewTrackingActive = isAutoViewTrackingActive;
+    CountlyViewTrackingInternal.sharedInstance.isAutoViewTrackingActive = isAutoViewTrackingActive;
 }
 
 - (BOOL)isAutoViewTrackingActive
 {
     CLY_LOG_I(@"%s", __FUNCTION__);
 
-    return CountlyViewTracking.sharedInstance.isAutoViewTrackingActive;
+    return CountlyViewTrackingInternal.sharedInstance.isAutoViewTrackingActive;
 }
 #endif
 
+- (CountlyViewTracking *) views
+{
+    return CountlyViewTracking.sharedInstance;
+}
 
 
 #pragma mark - User Details